packages/core: move sidebar things closer to the sidebar
